Mock BitbucketBambooUpdateService for start exercise test
